requirements.txt
appdirs==1.4.3
async-generator==1.10
beautifulsoup4==4.8.2
cachetools==4.2.2
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
Click==7.0
colorama==0.4.3
configobj==5.0.6
cryptography==37.0.4
debugpy==1.6.0
decorator==5.1.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
httplib2==0.14.0
importlib-metadata==4.12.0
importlib-resources==5.7.1
ipython==8.3.0
ipython-genutils==0.2.0
jsonschema==4.4.0
lxml==4.8.0
matplotlib==3.5.3
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==4.2.0
MouseInfo==0.1.3
newspaper3k==0.2.8
nltk==3.7
notebook==6.4.11
numpy==1.23.0
oauthlib==3.1.0
opencv-python==4.6.0.66
pandas==1.4.4
pandocfilters==1.5.0
Pillow==9.1.1
playsound==1.3.0
prompt-toolkit==3.0.29
psutil==5.5.1
PyAutoGUI==0.9.53
PyScreeze==0.1.28
pytesseract==0.3.9
python-telegram-bot==13.14
pytube==12.1.0
requests==2.28.1
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
sklearn==0.0
telebot==0.0.4
Telethon==1.24.0
tesseract==0.1.3
torch==1.12.0
tornado==6.1
typing_extensions==4.3.0
Unidecode==1.1.1
urllib3==1.26.10
youtube-dl==2021.4.26
python-dotenv==0.21.0